requirements.txt
uvicorn
fastapi
mangum
pydantic
psycopg2-binary
PyJWT
passlib